COMMERCE-1236 Refactor catalog dto converters
